refactor: separate MenuLayout from BaseLayout
