+++ src/main/kotlin/support/views/Components.kt
@@ -100,6 +100,7 @@ fun createNotification(text: String, durationValue: Int = 1000): Notification {
 
 fun createTabs(components: List<Component>): Component {
     return Tabs().apply {
+        setWidthFull()
         orientation = Tabs.Orientation.VERTICAL
         isAutoselect = false
         addThemeVariants(TabsVariant.LUMO_MINIMAL)
